feat: extract dataset manipulation from map features
