+++ src/pages/2_🌍_map.py
@@ -6,8 +6,6 @@ st.set_page_config(
     layout="wide",
 )
 
-from utils.st_logs import parse_log_buffer, init_logging_session_states
-
 from maps.obs_map import add_obs_map_header 
 from maps.alps_map import present_alps_map
 from maps.obs_map import present_obs_map
@@ -16,10 +14,6 @@ from datasets import disable_caching
 disable_caching()
 
 ############################################################
-# TO- DO: MAKE ENV FILE 
-# the dataset of observations (hf dataset in our space)
-dataset_id = "Saving-Willy/temp_dataset"
-data_files = "data/train-00000-of-00001.parquet"
 USE_BASIC_MAP = False
 DEV_SIDEBAR_LIB = True
 ############################################################
@@ -35,10 +29,7 @@ with tab_map_ui_cols[1]:
     
 if show_db_points:
     # show a nicer map, observations marked, tileset selectable.
-    st_observation = present_obs_map(
-        dataset_id=dataset_id, data_files=data_files,
-        dbg_show_extra=dbg_show_extra)
-    
+    st_observation = present_obs_map(dbg_show_extra=dbg_show_extra)
 else:
     # development map.
     st_observation = present_alps_map()
